Обернули все экраны во "Вью"
