--- js/logic/get-element-from-template.js
@@ -1,8 +1,0 @@
-
-export const main = document.querySelector(`main.central`);
-export const createScreen = (screen) => {
-  const element = document.createElement(`div`);
-  element.innerHTML = screen;
-  return element;
-};
-
