+++ js/page-parts/game-screens.js
@@ -1,5 +1,3 @@
-
-
 import {gameSets} from '../data/game-sets';
 import gameState from '../data/main-data';
