--- js/screens/stats.js
@@ -1,117 +1,0 @@
-import headerTemplate from '../page-parts/header';
-import getBack from '../logic/get-back-to-start-btn.js';
-import gameState from '../data/main-data';
-import gameAnswers from '../page-parts/answers-list';
-import footer from '../page-parts/footer';
-// import {main} from '../logic/units';
-const screenGameStats = () => {
-  const markup = `
-  ${headerTemplate.empty}
-<div class="result">
-<h1>${gameState.status}</h1>
-<table class="result__table">
-  <tr>
-    <td class="result__number">1.</td>
-    <td colspan="2">
-      <ul class="stats">
-        ${gameAnswers(gameState.answers)}
-      </ul>
-    </td>
-    <td class="result__points">×&nbsp;100</td>
-    <td class="result__total">900</td>
-  </tr>
-  <tr>
-    <td></td>
-    <td class="result__extra">Бонус за скорость:</td>
-    <td class="result__extra">1&nbsp;<span class="stats__result stats__result--fast"></span></td>
-    <td class="result__points">×&nbsp;50</td>
-    <td class="result__total">50</td>
-  </tr>
-  <tr>
-    <td></td>
-    <td class="result__extra">Бонус за жизни:</td>
-    <td class="result__extra">2&nbsp;<span class="stats__result stats__result--alive"></span></td>
-    <td class="result__points">×&nbsp;50</td>
-    <td class="result__total">100</td>
-  </tr>
-  <tr>
-    <td></td>
-    <td class="result__extra">Штраф за медлительность:</td>
-    <td class="result__extra">2&nbsp;<span class="stats__result stats__result--slow"></span></td>
-    <td class="result__points">×&nbsp;50</td>
-    <td class="result__total">-100</td>
-  </tr>
-  <tr>
-    <td colspan="5" class="result__total  result__total--final">950</td>
-  </tr>
-</table>
-<table class="result__table">
-  <tr>
-    <td class="result__number">2.</td>
-    <td>
-      <ul class="stats">
-        <li class="stats__result stats__result--wrong"></li>
-        <li class="stats__result stats__result--slow"></li>
-        <li class="stats__result stats__result--fast"></li>
-        <li class="stats__result stats__result--correct"></li>
-        <li class="stats__result stats__result--wrong"></li>
-        <li class="stats__result stats__result--unknown"></li>
-        <li class="stats__result stats__result--slow"></li>
-        <li class="stats__result stats__result--wrong"></li>
-        <li class="stats__result stats__result--fast"></li>
-        <li class="stats__result stats__result--wrong"></li>
-      </ul>
-    </td>
-    <td class="result__total"></td>
-    <td class="result__total  result__total--final">fail</td>
-  </tr>
-</table>
-<table class="result__table">
-  <tr>
-    <td class="result__number">3.</td>
-    <td colspan="2">
-      <ul class="stats">
-        <li class="stats__result stats__result--wrong"></li>
-        <li class="stats__result stats__result--slow"></li>
-        <li class="stats__result stats__result--fast"></li>
-        <li class="stats__result stats__result--correct"></li>
-        <li class="stats__result stats__result--wrong"></li>
-        <li class="stats__result stats__result--unknown"></li>
-        <li class="stats__result stats__result--slow"></li>
-        <li class="stats__result stats__result--unknown"></li>
-        <li class="stats__result stats__result--fast"></li>
-        <li class="stats__result stats__result--unknown"></li>
-      </ul>
-    </td>
-    <td class="result__points">×&nbsp;100</td>
-    <td class="result__total">900</td>
-  </tr>
-  <tr>
-    <td></td>
-    <td class="result__extra">Бонус за жизни:</td>
-    <td class="result__extra">2&nbsp;<span class="stats__result stats__result--alive"></span></td>
-    <td class="result__points">×&nbsp;50</td>
-    <td class="result__total">100</td>
-  </tr>
-  <tr>
-    <td colspan="5" class="result__total  result__total--final">950</td>
-  </tr>
-</table>
-</div>
-${footer}`;
-  return markup;
-};
-const main = document.querySelector(`main.central`);
-
-const renderStats = () => {
-  main.innerHTML = ``;
-  const stats = document.createElement(`div`);
-  stats.innerHTML = screenGameStats();
-  main.appendChild(stats);
-  getBack(`.back`, main);
-
-
-};
-
-
-export default renderStats;
